code/deployment/api/requirements.txt
fastapi
uvicorn
torch
pydantic
numpy
Pillow
python-multipart